Move buffer utilities into helpers.go
