+++ runtime/tchannel/helpers_test.go
@@ -22,6 +22,8 @@ package tchannel
 
 import (
 	"bytes"
+	"math/rand"
+	"sync"
 	"testing"
 
 	"github.com/uber/tchannel-go/testutils/testreader"
@@ -51,4 +53,29 @@ func TestEnsureEmptyError(t *testing.T) {
 	err := EnsureEmpty(reader, "has bytes")
 	require.Error(t, err, "ensureEmpty should fail when there's an error")
 	assert.Equal(t, testreader.ErrUser, err, "Unexpected error")
+}
+
+func TestBuffers(t *testing.T) {
+	var wg sync.WaitGroup
+	for g := 0; g < 10; g++ {
+		wg.Add(1)
+		go func() {
+			for i := 0; i < 100; i++ {
+				buf := GetBuffer()
+				assert.Zero(t, buf.Len(), "Expected truncated buffer")
+
+				bytesOfNoise := make([]byte, rand.Intn(5000))
+				_, err := rand.Read(bytesOfNoise)
+				assert.NoError(t, err, "Unexpected error from rand.Read")
+				_, err = buf.Write(bytesOfNoise)
+				assert.NoError(t, err, "Unexpected error from buffer.Write")
+
+				assert.Equal(t, buf.Len(), len(bytesOfNoise), "Expected same buffer size")
+
+				PutBuffer(buf)
+			}
+			wg.Done()
+		}()
+	}
+	wg.Wait()
 }
